Final catalyst-dl trace version
- Some docs written
- trace_model is in utils now
- Some catalyst.dl.experiements module structure refactoring
