+++ catalyst/dl/__main__.py
@@ -2,9 +2,15 @@ from argparse import ArgumentParser, RawTextHelpFormatter
 from collections import OrderedDict
 
 from catalyst.__version__ import __version__
-from .scripts import run, make_report
+from .scripts import trace, run, make_report
 
-COMMANDS = OrderedDict([("run", run), ("make-report", make_report)])
+COMMANDS = OrderedDict(
+    [
+        ("run", run),
+        ("trace", trace),
+        ("make-report", make_report),
+    ]
+)
 
 
 def build_parser() -> ArgumentParser:
