+++ catalyst/dl/utils/__init__.py
@@ -1,0 +1,15 @@
+from .utils import \
+    UtilsFactory, \
+    get_activation_by_name, \
+    get_optimizer_momentum, \
+    set_optimizer_momentum, \
+    prepare_optimizable_params, \
+    assert_fp16_available
+
+from .trace import trace_model
+
+__all__ = [
+    "UtilsFactory", "get_activation_by_name", "set_optimizer_momentum",
+    "get_optimizer_momentum", "prepare_optimizable_params",
+    "assert_fp16_available", "trace_model"
+]
